Refactoring Validator and Reflections to modules.
