+++ src/main/java/br/com/ppm/commons/Strings.java
@@ -16,6 +16,8 @@
  */
 package br.com.ppm.commons;
 
+import br.com.ppm.commons.validation.Validator;
+
 /**
  * String Operations
  *
